[Create] ProductModel 관리를 위해서 ProductListModel 생성
